+++ apps/web/src/lib/url.ts
@@ -1,0 +1,11 @@
+import type { ReadonlyURLSearchParams } from "next/navigation";
+
+export const createUrl = (
+  pathname: string,
+  params: URLSearchParams | ReadonlyURLSearchParams
+) => {
+  const paramsString = params.toString();
+  const queryString = `${paramsString.length ? "?" : ""}${paramsString}`;
+
+  return `${pathname}${queryString}`;
+};
